feat(openai): add HTTPX async adapter support
