+++ src/llm_api_adapter/adapters/openai/__init__.py
@@ -1,0 +1,5 @@
+"""OpenAI adapter package."""
+
+from .adapter import OpenAIAdapter
+
+__all__ = ["OpenAIAdapter"]
